# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Family Manager BI Group API",
        "version": "1.1.0"
    },
    "paths": {
...
        "/api/cases/{id}": {
            "get": {
                "tags": [
                    "Cases"
                ],
                "description": "Get single case",
                "operationId": "App\\Http\\Controllers\\CaseController::show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Case id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return category data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CaseSearchResource"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "USER_HAS_NO_PERMISSION"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Cases"
                ],
                "description": "Update case",
                "operationId": "App\\Http\\Controllers\\CaseController::update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Case id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CaseRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cases sucsessfully updates",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CaseUpdateRequest"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "USER_HAS_NO_PERMISSION"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Cases"
                ],
                "description": "Delete case",
                "operationId": "App\\Http\\Controllers\\CaseController::destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Case id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Case successfully deleted"
                    },
                    "403": {
                        "description": "USER_HAS_NO_PERMISSION"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/categories/{id}": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "description": "Get single category",
                "operationId": "App\\Http\\Controllers\\CategoryController::show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Category id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return category data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CategoryRequest"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "USER_HAS_NO_PERMISSION"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Categories"
                ],
                "description": "Update category",
                "operationId": "App\\Http\\Controllers\\CategoryController::update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Category id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CategoryRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Category sucsessfully updated"
                    },
                    "403": {
                        "description": "USER_HAS_NO_PERMISSION"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Categories"
                ],
                "description": "Delete category",
                "operationId": "App\\Http\\Controllers\\CategoryController::destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Category id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category successfully deleted"
                    },
                    "403": {
                        "description": "USER_HAS_NO_PERMISSION"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/family-compositions": {
            "get": {
                "tags": [
                    "Family Compositions"
                ],
                "description": "Get list of family compositions with search",
                "operationId": "App\\Http\\Controllers\\FamilyCompositionController::index",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search string",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "семейств"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "List limit",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "12"
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "description": "Offset",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "0"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return compositions list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FamilyCompositionRequest"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "USER_HAS_NO_PERMISSION"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Family Compositions"
                ],
                "operationId": "App\\Http\\Controllers\\FamilyCompositionController::store",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/FamilyCompositionRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Return bool"
                    },
                    "403": {
                        "description": "USER_HAS_NO_PERMISSION"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/family-compositions/{id}": {
            "get": {
                "tags": [
                    "Family Compositions"
                ],
                "description": "Get single family composition",
                "operationId": "App\\Http\\Controllers\\FamilyCompositionController::show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Composition id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return family composition data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FamilyCompositionRequest"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "USER_HAS_NO_PERMISSION"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Family Compositions"
                ],
                "description": "Update family composition",
                "operationId": "App\\Http\\Controllers\\FamilyCompositionController::update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Composition id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/FamilyCompositionRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Return updated family composition"
                    },
                    "403": {
                        "description": "USER_HAS_NO_PERMISSION"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Family Compositions"
                ],
                "description": "Delete family composition",
                "operationId": "App\\Http\\Controllers\\FamilyCompositionController::destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Composition id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Family composition Successfully deleted"
                    },
                    "403": {
                        "description": "USER_HAS_NO_PERMISSION"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/reports/{id}": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "description": "Get single report",
                "operationId": "App\\Http\\Controllers\\ReportController::show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Report id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return report resource",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ReportResource"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "USER_HAS_NO_PERMISSION"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Reports"
                ],
                "description": "Delete report",
                "operationId": "App\\Http\\Controllers\\ReportController::destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Report id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Report successfully deleted"
                    },
                    "403": {
                        "description": "USER_HAS_NO_PERMISSION"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/sections/{id}": {
            "get": {
                "tags": [
                    "Sections"
                ],
                "description": "Get single section",
                "operationId": "App\\Http\\Controllers\\SectionController::show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Section id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return section data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SectionRequest"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "USER_HAS_NO_PERMISSION"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Sections"
                ],
                "description": "Update section",
                "operationId": "App\\Http\\Controllers\\SectionController::update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Section id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SectionRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Return updated section"
                    },
                    "403": {
                        "description": "USER_HAS_NO_PERMISSION"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Sections"
                ],
                "description": "Delete section",
                "operationId": "App\\Http\\Controllers\\SectionController::destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Section id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Section successfully deleted"
                    },
                    "403": {
                        "description": "USER_HAS_NO_PERMISSION"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/task-types/{id}": {
            "get": {
                "tags": [
                    "Tasks Types"
                ],
                "description": "Get single task type",
                "operationId": "App\\Http\\Controllers\\TaskTypeController::show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Task id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return task type data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TaskTypeRequest"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "USER_HAS_NO_PERMISSION"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Tasks Types"
                ],
                "description": "Update task type",
                "operationId": "App\\Http\\Controllers\\TaskTypeController::update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Task id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TaskTypeRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Return updated task type",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TaskTypeRequest"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "USER_HAS_NO_PERMISSION"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Tasks Types"
                ],
                "description": "Delete task type",
                "operationId": "App\\Http\\Controllers\\TaskTypeController::destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Task id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task successfully deleted"
                    },
                    "403": {
                        "description": "USER_HAS_NO_PERMISSION"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CaseRequest": {
                "title": "Case request",
                "description": "Case request body data",
                "required": [
                    "name"
                ],
                "properties": {
                    "CaseName": {
                        "schema": "CaseRequest",
                        "type": "string",
                        "example": "(Dr)_ Door78_Дверь Двустворчатая_Внутренняя_ГОСТ 31173"
                    },
                    "CaseHashAfter": {
                        "schema": "CaseRequest",
                        "type": "string",
                        "example": "a35bc872-1910-11eb-950d-00155d34d6e2"
                    },
                    "CaseHashBefore": {
                        "schema": "CaseRequest",
                        "type": "string",
                        "example": "a35bc872-1910-11eb-950d-00155d34d6e2"
                    },
                    "CategoryID": {
                        "schema": "CaseRequest",
                        "type": "integer",
                        "example": 9
                    },
                    "SectionID": {
                        "schema": "CaseRequest",
                        "type": "integer",
                        "example": 1
                    },
                    "Files": {
                        "schema": "CaseRequest",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/CaseFileRequest"
                        }
                    }
                },
                "type": "object"
            },
            "CaseUpdateRequest": {
                "title": "Case update request",
                "description": "Case update request body data",
                "required": [
                    "name, section_id, category_id"
                ],
                "properties": {
                    "name": {
                        "schema": "CaseUpdateRequest",
                        "type": "string"
                    },
                    "section_id": {
                        "schema": "CaseUpdateRequest",
                        "type": "integer"
                    },
                    "category_id": {
                        "schema": "CaseUpdateRequest",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        []
    ]
}